requirements.txt



Pillow == 8.3.1


keras == 2.6.0


numpy == 1.20.3


opencv_python == 4.4.0.46

streamlit == 0.85.0


tensorflow == 2.6.0
